Correcting unexpected behaviour on the --isLatino parser argument
When you use type=bool, argparse treats the string "False" as truthy, getting converted to True. This is because any non-empty string in Python is considered truthy (including "False"), and bool("False") results in True. Thus, --isLatino cannot be set to False.

To correct this issue, adding a function to parse boolean-like strings and then pass it to argparse.